backend/requirements.txt
Django==3.2.*
celery==5.1.*
pika==1.3.2
psycopg2==2.9.6
elasticsearch==8.6.1
djangorestframework==3.12.2
django-cors-headers==3.3.0
django-cors-middleware==1.3.1